requirements.txt
aliyun-python-sdk-core-v3==2.8.9
asn1crypto==0.24.0
Babel==2.6.0
blinker==1.4
blueprint==3.4.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
cryptography==2.3
decorator==4.3.0
defusedxml==0.5.0
Django==1.8
django-grappelli==2.10.1
dominate==2.3.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Moment==0.6.0
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
guess-language==0.2
gunicorn==19.9.0
idna==2.6
itsdangerous==0.24
jenkins==1.0.2
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
pbr==4.1.1
pycparser==2.18
pycryptodome==3.6.4
PyMySQL==0.9.2
python3-openid==3.1.0
pytz==2018.5
requests==2.18.4
selenium==3.8.1
six==1.11.0
SQLAlchemy==1.2.10
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
urllib3==1.22
uWSGI==2.0.15
visitor==0.1.3
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.2.1
